// src/app/models/employee/Employee.model.ts
export interface Employee {
    id: number;
    employeeCode: number;
    employeeName: string;
    employeeRole: string;
    employeeLevel: string;
    employeeDepartment: string;
    birthday: string;
    idCard: string;
    salary: string;
    phoneNumber: string;
    email: string;
    address: string;
    gender: string;
}